move cecil usage around
