+++ src/Integration/Directory.Build.props
@@ -1,1 +1,1 @@
-<Project><PropertyGroup><IntegrationVersion>3.0.0-packaging0001</IntegrationVersion></PropertyGroup></Project>
+<Project><PropertyGroup><IntegrationVersion>3.0.0-fixpackaging0001</IntegrationVersion></PropertyGroup></Project>
