+++ src/ScriptBuilder/AttributeReading/Settings.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using NServiceBus.Persistence.Sql.ScriptBuilder;
 
-class Settings
+public class Settings
 {
     public List<BuildSqlVariant> BuildVariants;
     public string ScriptPromotionPath;
