+++ src/ScriptBuilder/Writers/OutboxWriter.cs
