+++ src/ScriptBuilder/Writers/SagaWriter.cs
