+++ src/ScriptBuilder/Writers/SubscriptionWriter.cs
