+++ src/ScriptBuilderTask/InnerTask.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.IO;
-using Mono.Cecil;
 using NServiceBus.Persistence.Sql;
 
 class InnerTask
@@ -22,31 +21,9 @@ class InnerTask
 
     public void Execute()
     {
-        var module = ModuleDefinition.ReadModule(assemblyPath, new ReaderParameters(ReadingMode.Deferred));
         var scriptPath = Path.Combine(intermediateDirectory, "NServiceBus.Persistence.Sql");
         DirectoryExtensions.Delete(scriptPath);
-        var settings = SettingsAttributeReader.Read(module);
-        foreach (var variant in settings.BuildVariants)
-        {
-            var variantPath = Path.Combine(scriptPath, variant.ToString());
-            Directory.CreateDirectory(variantPath);
-            if (settings.ProduceSagaScripts)
-            {
-                SagaWriter.WriteSagaScripts(variantPath, module, variant, logError);
-            }
-            if (settings.ProduceTimeoutScripts)
-            {
-                TimeoutWriter.WriteTimeoutScript(variantPath, variant);
-            }
-            if (settings.ProduceSubscriptionScripts)
-            {
-                SubscriptionWriter.WriteSubscriptionScript(variantPath, variant);
-            }
-            if (settings.ProduceOutboxScripts)
-            {
-                OutboxWriter.WriteOutboxScript(variantPath, variant);
-            }
-        }
+        var settings = ScriptWriter.Write(assemblyPath, scriptPath, logError);
 
         PromoteFiles(scriptPath, settings);
     }
