+++ src/ScriptBuilderTask/ScriptBuilderTask.csproj
@@ -22,7 +22,6 @@
   <ItemGroup>
     <PackageReference Include="GitVersionTask" Version="4.0.0-beta0012" PrivateAssets="All" />
     <PackageReference Include="ILRepack" Version="2.0.13" PrivateAssets="All" />
-    <PackageReference Include="Mono.Cecil" Version="0.10.0-beta5" PrivateAssets="All" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net452' ">
